projects/AutoML Copilot/requirements.txt
streamlit>=1.24.0
pandas>=1.5
numpy>=1.23
plotly>=5.6
matplotlib>=3.5
seaborn>=0.12
scikit-learn>=1.1
python-dotenv>=1.0
requests>=2.28
openai>=1.40
# Optional AutoML (heavy). Install only if you have resources:
# pycaret>=3.0
tqdm>=4.64
